Перед добавлением отчетов в EXEL

+++ src/main/java/quartztop/analitics/dtos/reports/StockByStore.java
@@ -13,9 +13,9 @@ import java.util.regex.Pattern;
 public class StockByStore {
     private Meta meta;
     private String name;
-    private int stock;
-    private int reserve;
-    private int inTransit;
+    private float stock;
+    private float reserve;
+    private float inTransit;
     private UUID productId;
     private UUID storeId;
